+++ src/analyzers/age/analyzer.rs
@@ -6,8 +6,8 @@ use crate::TextItem;
 use super::data::WEIGHTS_AGE;
 
 pub struct Age {
-    pub items: HashMap<String, f32>,
-    pub intercept: f32,
+    pub items: HashMap<String, f64>,
+    pub intercept: f64,
 }
 
 impl Age {
@@ -27,15 +27,15 @@ impl Age {
     }
 
     #[inline(always)]
-    pub fn get_entry(&self, term: &str) -> Option<&f32> {
+    pub fn get_entry(&self, term: &str) -> Option<&f64> {
         self.items.get(term)
     }
 
     #[inline(always)]
-    pub fn get_score(&self, item: &TextItem, word: &str) -> Option<f32> {
+    pub fn get_score(&self, item: &TextItem, word: &str) -> Option<f64> {
         let word_coeff =
-            ((*item.word_freqs.get(word)?) as f32)
-                .div(item.word_count as f32)
+            ((*item.word_freqs.get(word)?) as f64)
+                .div(item.word_count as f64)
                 .mul(self.get_entry(word)?);
 
         Some(
@@ -44,11 +44,11 @@ impl Age {
     }
 
     #[inline(always)]
-    pub fn run(&self, item: &TextItem) -> f32 {
+    pub fn run(&self, item: &TextItem) -> f64 {
         item.word_freqs
             .keys()
             .filter_map(|x| self.get_score(item, x))
-            .sum::<f32>()
+            .sum::<f64>()
             + self.intercept
     }
 }
