+++ src/types/rated.rs
@@ -3,5 +3,5 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize, Debug)]
 pub struct LexiconRated {
     pub term: String,
-    pub rating: f32,
+    pub rating: f64,
 }
